pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "budsctl"
version = "0.1.0"
description = "Bluetooth earbuds control with pluggable YAML protocols"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
  "pyyaml>=6.0.1",
  "jsonschema>=4.23.0",
  "typer>=0.12.3",
  "bleak>=0.22.3",
]

[project.scripts]
budsctl = "budsctl.cli:run"

[tool.hatch.build.targets.wheel]
packages = ["budsctl"]

[tool.hatch.build]
include = [
  "budsctl/plugins/*.yaml",
  "budsctl/schemas/*.json",
]

[dependency-groups]
dev = [
  "pytest>=8.2.0",
]
